requirements.txt
Flask>=2.3.0
langgraph>=0.0.60
langchain>=0.1.0
langchain-openai>=0.0.5
openai>=1.6.0
PyPDF2>=3.0.0
pdfplumber>=0.9.0
python-dotenv>=1.0.0
Werkzeug>=2.3.0
Jinja2>=3.1.0
MarkupSafe>=2.1.0
itsdangerous>=2.1.0
click>=8.1.0